move interface to shared, add id to result
