+++ shared/src/common/types/top-autoline/top-autoline.ts
@@ -1,0 +1,17 @@
+interface TopCar {
+  id: string;
+  name: string;
+  url: string;
+  photoUrl: string;
+  brand: {
+    name: string;
+    logoUrl: string;
+  };
+  price: number;
+  race: number;
+  transmission: string;
+  location: string;
+  fuelType: string;
+}
+
+export type { TopCar };
